# Optimize parsing option args in `JSON.parse`